mercenary behaving as expected, battles presumably working:

+++ src/main/java/dungeonmania/interfaces/Health.java
@@ -2,10 +2,10 @@ package dungeonmania.interfaces;
 
 public interface Health {
     /*Same as Enemy */
-    public int getHealth();
-    public void setHealth(int health);
-    public int getAttackDamage();
-    public void loseHealth(int deltaHealth);
+    public double getHealth();
+    public void setHealth(double health);
+    public double getAttackDamage();
+    public void loseHealth(double deltaHealth);
     public int getEntityId();
     public String getType();
 }
